Move more of the videosource code to the videosourcewrappers file (#4366)
